migrate lettings to new app lettings
