+++ lettings/templates/lettings/lettings_index.html
@@ -4,7 +4,7 @@
     <ul>
     {% for letting in lettings_list %}
         <li>
-            <a href="{% url 'letting' letting_id=letting.id %}">
+            <a href="{% url 'lettings:letting' letting_id=letting.id %}">
                 {{ letting.title }}
             </a>
         </li>
