+++ oc_lettings_site/settings.py
@@ -26,6 +26,7 @@ INSTALLED_APPS = [
     'django.contrib.sessions',
     'django.contrib.messages',
     'django.contrib.staticfiles',
+    'lettings',
 ]
 
 MIDDLEWARE = [
