+++ oc_lettings_site/views.py
@@ -1,5 +1,5 @@
 from django.shortcuts import render
-from .models import Letting, Profile
+from .models import Profile
 
 
 """ Lorem ipsum dolor sit amet, consectetur adipiscing elit.
@@ -14,40 +14,6 @@ Praesent volutpat porttitor magna, non finibus neque cursus id."""
 
 def index(request):
     return render(request, 'index.html')
-
-
-""" Aenean leo magna, vestibulum et tincidunt fermentum, consectetur quis velit.
-Sed non placerat massa. Integer est nunc, pulvinar a tempor et, bibendum id arcu.
-Vestibulum ante ipsum primis in faucibus orci luctus et ultrices posuere cubilia curae;
-Cras eget scelerisque """
-
-
-def lettings_index(request):
-    lettings_list = Letting.objects.all()
-    context = {'lettings_list': lettings_list}
-    return render(request, 'lettings_index.html', context)
-
-
-""" Cras ultricies dignissim purus, vitae hendrerit ex varius non.
-In accumsan porta nisl id eleifend. Praesent dignissim, odio eu consequat pretium,
-purus urna vulputate arcu, vitae efficiturlacus justo nec purus.
-Aenean finibus faucibus lectus at porta. Maecenas auctor, est ut luctus congue,
-dui enim mattis enim, ac condimentum velit libero in magna. Suspendisse potenti.
-In tempus a nisi sed laoreet. Suspendisse porta dui eget sem accumsan interdum.
-Ut quis urna pellentesque justo mattis ullamcorper ac non tellus.
-In tristique mauris eu velit fermentum, tempus pharetra est luctus.
-Vivamus consequat aliquam libero, eget bibendum lorem. Sed non dolor risus.
-Mauris condimentum auctor elementum. Donec quis nisi ligula.
-Integer vehicula tincidunt enim, ac lacinia augue pulvinar sit amet. """
-
-
-def letting(request, letting_id):
-    letting = Letting.objects.get(id=letting_id)
-    context = {
-        'title': letting.title,
-        'address': letting.address,
-    }
-    return render(request, 'letting.html', context)
 
 
 """ Sed placerat quam in pulvinar commodo. Nullam laoreet consectetur ex,
