+++ templates/profile.html
@@ -6,4 +6,4 @@
 <p>Favorite city: {{ profile.favorite_city }}</p>
 <div><a href="{% url 'profiles_index' %}">Back</a></div>
 <div><a href="{% url 'index' %}">Home</a></div>
-<div><a href="{% url 'lettings_index' %}">Lettings</a></div>
+<div><a href="{% url 'lettings:lettings_index' %}">Lettings</a></div>
