+++ templates/profiles_index.html
@@ -14,4 +14,4 @@
     <p>No profiles are available.</p>
 {% endif %}
 <div><a href="{% url 'index' %}">Home</a></div>
-<div><a href="{% url 'lettings_index' %}">Lettings</a></div>
+<div><a href="{% url 'lettings:lettings_index' %}">Lettings</a></div>
